test refactor wip
